requirements.txt
accelerate==1.7.0
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==25.3.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.2.1
colorama==0.4.6
corsheaders==4.7.0
cryptography==45.0.3
Django==5.2
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
distro==1.9.0
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.6.0
fsspec==2025.5.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
huggingface-hub==0.32.4
idna==3.4
jinja2==3.1.4
joblib==1.5.1
langdetect==1.0.9
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.4
mysql-connector-python==8.0.33
mysqlclient==2.2.7
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.3.0
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==11.2.1
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydotplus==2.0.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.7
requests==2.28.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.52.4
tzdata==2025.2
typing-extensions==4.12.2
typing-inspection==0.4.1
urllib3==1.26.13
uvicorn==0.23.2
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0 